proxy/conntrack: moved pkg/util/conntrack -> pkg/proxy/conntrack
