--- pkg/util/conntrack/OWNERS
@@ -1,8 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-approvers:
-  - sig-network-approvers
-reviewers:
-  - sig-network-reviewers
-labels:
-  - sig/network
